Model test related methods

BiochemNetABC.change_simulation_stop_criteria — Method

change_simulation_stop_criteria(m::ContinuousTimeModel, isabsorbing_func::Symbol)

Change the simulation of the model m by adding a stop criteria based on the function named isabsorbing_func::Symbol. isabsorbing_func must have the type signature isabsorbing_func(p::Vector{Float64}, x::Vector{Int}) where p is the parameter vector of the model and x a state (not an observed state) of the model.

BiochemNetABC.distribute_mean_value_lha — Method

distribute_mean_value_lha(sm::SynchronizedModel, sym_var::Symbol, nbr_stim::Int)

Distribute over workers the computation of the mean value of an LHA over nbr_sim simulations of the model.

BiochemNetABC.draw! — Method

draw!(mat_p, pm)

Draw size(mat_p)[2] (number of columns of mat_p) parameters from the prior distribution defined in pm and stores it in mat_p.

BiochemNetABC.draw! — Method

draw!(vec_p, pm)

Draw a parameter from the prior distribution defined in pm and stores it in vec_p.

BiochemNetABC.draw_model! — Method

draw_model!(pm::ParametricModel)

Draw a parameter from the prior disitribution defined in pm::ParametricModel and save it in the model contained in pm.

BiochemNetABC.prior_pdf! — Method

prior_pdf(res_pdf, mat_p, pm)

Computes the density for each column of mat_p of the prior distribution defined in pm. Stores it in res_pdf. (length(vec_res) == size(mat_p)[2])

BiochemNetABC.prior_pdf — Method

prior_pdf(p_prior, pm)

Computes the density at p_prior of the prior distribution defined in pm.

BiochemNetABC.probability_var_value_lha — Method

distribute_var_value_lha(sm::SynchronizedModel, nbr_sim::Int, value = 0, sym_var = :d)

Compute the probability that the variable sym_var is equal to value of an LHA over nbr_sim simulations of the model.

BiochemNetABC.simulate — Method

simulate(pm::ParametricModel, p_prior::AbstractVector{Float64})

Simulates the model contained in the parametric model pm with the p_prior parameters. It simulates the model by taking the parameters contained in get_proba_model(pm).p and replace the 1D parameters pm.params with p_prior.

BiochemNetABC.volatile_simulate — Method

volatile_simulate(pm::ParametricModel, p_prior::AbstractVector{Float64})

A volatile version of simulate(pm::ParametricModel, p_prior::AbstractVector{Float64}). The model in pm should be of type SynchronizedModel (typeof(pm.m) <: SynchronizedModel). It returns S::StateLHA, not a trajectory.

BiochemNetABC.volatile_simulate — Method

volatile_simulate(sm::SynchronizedModel; p, verbose)

Simulates a model synchronized with an automaton but does not store the values of the simulation in order to improve performance. It returns the last state of the simulation S::StateLHA not a trajectory σ::SynchronizedTrajectory.